Add theme picker tests
