+++ package.json
@@ -15,6 +15,7 @@
   },
   "dependencies": {
     "@testing-library/react": "^11.2.7",
+    "@testing-library/user-event": "^13.1.9",
     "jest-axe": "^4.1.0",
     "lz-string": "^1.4.4",
     "next": "^10.2.3",
