+++ page-tests/index.test.tsx
@@ -30,16 +30,12 @@ describe(Page.name, () => {
 
   it("should render without accessibility errors when no message", async () => {
     const encodedMessage = "";
-    const currentUrl = "https://example.com/";
-    const host = "https://example.com";
 
     const page = render(
       <MessageStore>
         <LanguageStore>
           <Page
             encodedParamMessage={encodedMessage}
-            currentUrl={currentUrl}
-            host={host}
           />
         </LanguageStore>
       </MessageStore>,
